% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\name{import_flights}
\alias{import_flights}
\alias{prepare_flights}
\alias{download_flights}
\alias{download_carriers}
\alias{import_carriers}
\alias{load_flights}
\title{Import flights}
\usage{
import_flights(years = 2024:2025, months = 1:12)

prepare_flights(files = NULL, years = NULL, months = NULL)

download_flights(years = 2025, months = 1:12)

download_carriers()

import_carriers()

load_flights(files)
}
\arguments{
\item{years}{years integer vector}

\item{months}{months integer vector}

\item{files}{flights files}
}
\description{
Import flights

Prepare flights

Download flights

Download carriers

Import carriers

Load flights
}
