% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My30DayChartChallenge2026-package.R
\docType{package}
\name{My30DayChartChallenge2026-package}
\alias{My30DayChartChallenge2026}
\alias{My30DayChartChallenge2026-package}
\title{My30DayChartChallenge2026: My 30 Day Chart Challenge 2026}
\description{
One plot for 30 days.
}
\author{
\strong{Maintainer}: Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID}) [copyright holder]

}
\keyword{internal}
